Move authenticated fetch wrapper into app/api.
