+++ client/src/app/api/authFetch.ts
