+++ client/src/app/api/transactions/route.ts
@@ -1,6 +1,6 @@
 import { getServerSession } from 'next-auth';
 import { authOptions } from '@/app/api/auth/[...nextauth]/route';
-import { authFetch } from '@/lib/authFetch';
+import { authFetch } from '@/app/api/authFetch';
 
 export async function GET() {
     const session = await getServerSession(authOptions);
